requirements.txt
fastapi==0.115.8
python-dotenv==1.0.1
uvicorn==0.34.0
openai==1.65.5
httpx==0.28.1
pydantic==2.10.6
python-multipart==0.0.20
starlette==0.45.3